+++ testing/tests/DevExpress.ui.widgets.dataGrid/dataGrid.tests.js
@@ -67,11 +67,9 @@ import { setTemplateEngine } from 'core/templates/template_engine_registry';
 import gridCoreUtils from 'ui/grid_core/ui.grid_core.utils';
 import fx from 'animation/fx';
 import config from 'core/config';
-import keyboardMock from '../../helpers/keyboardMock.js';
 import pointerMock from '../../helpers/pointerMock.js';
 import pointerEvents from 'events/pointer';
 import ajaxMock from '../../helpers/ajaxMock.js';
-import themes from 'ui/themes';
 import DataGridWrapper from '../../helpers/wrappers/dataGridWrappers.js';
 import { checkDxFontIcon, DX_ICON_XLSX_FILE_CONTENT_CODE, DX_ICON_EXPORT_SELECTED_CONTENT_CODE } from '../../helpers/checkDxFontIconHelper.js';
 import 'ui/scroll_view';
@@ -79,7 +77,6 @@ import { createDataGrid, baseModuleConfig } from '../../helpers/dataGridHelper.j
 
 
 const DX_STATE_HOVER_CLASS = 'dx-state-hover';
-const TEXTEDITOR_INPUT_SELECTOR = '.dx-texteditor-input';
 const CELL_UPDATED_CLASS = 'dx-datagrid-cell-updated-animation';
 const ROW_INSERTED_CLASS = 'dx-datagrid-row-inserted-animation';
 const dataGridWrapper = new DataGridWrapper('#dataGrid');
@@ -132,27 +129,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
     QUnit.test('formatValue for grouped column with calculateGroupValue', function(assert) {
         assert.strictEqual(gridCore.formatValue('2012', { format: 'shortDate' }), '2012');
-    });
-
-    QUnit.test('Accessibility columns id should not set for columns editors (T710132)', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            columns: ['field1', 'field2'],
-            filterRow: { visible: true },
-            headerFilter: { visible: true },
-            searchPanel: { visible: true },
-            editing: { mode: 'row', allowUpdating: true },
-            dataSource: [{ field1: '1', field2: '2' }]
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.editRow(0);
-        this.clock.tick();
-
-        // assert
-        assert.equal($('.dx-texteditor [id]').length, 0, 'editors has no accessibility id');
     });
 
     QUnit.test('DataGrid - Should hide filter row menu after losing it\'s focus', function(assert) {
@@ -464,142 +440,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         });
     });
 
-    QUnit.test('Command column accessibility structure', function(assert) {
-    // arrange
-        createDataGrid({
-            columns: ['field1', 'field2'],
-            editing: { mode: 'row', allowAdding: true }
-        });
-
-        // assert
-        assert.equal($('.dx-row.dx-header-row').eq(0).attr('role'), 'row');
-        assert.equal($('.dx-header-row .dx-command-edit').eq(0).attr('role'), 'columnheader');
-        assert.equal($('.dx-header-row .dx-command-edit').eq(0).attr('aria-colindex'), 3);
-    });
-
-    QUnit.test('Command buttons should contains aria-label accessibility attribute if rendered as icons (T755185)', function(assert) {
-    // arrange
-        const columnsWrapper = dataGridWrapper.columns;
-        const dataGrid = createDataGrid({
-            dataSource: [{ id: 0, c0: 'c0' }],
-            columns: [
-                {
-                    type: 'buttons',
-                    buttons: ['edit', 'delete', 'save', 'cancel']
-                },
-                'id'
-            ],
-            editing: {
-                allowUpdating: true,
-                allowDeleting: true,
-                useIcons: true
-            }
-        });
-
-        this.clock.tick();
-
-        // assert
-        columnsWrapper.getCommandButtons().each((_, button) => {
-            const ariaLabel = $(button).attr('aria-label');
-            assert.ok(ariaLabel && ariaLabel.length, `aria-label '${ariaLabel}'`);
-        });
-
-        // act
-        dataGrid.editRow(0);
-        // assert
-        columnsWrapper.getCommandButtons().each((_, button) => {
-            const ariaLabel = $(button).attr('aria-label');
-            assert.ok(ariaLabel && ariaLabel.length, `aria-label '${ariaLabel}'`);
-        });
-    });
-
-
-    [false, true].forEach((useIcons) => {
-        QUnit.test(`Command buttons should be rendered with RTL when useIcons=${useIcons} (T915926)`, function(assert) {
-            // arrange
-            const columnsWrapper = dataGridWrapper.columns;
-            const dataGrid = createDataGrid({
-                rtlEnabled: true,
-                dataSource: [{ id: 0, c0: 'c0' }],
-                editing: {
-                    allowUpdating: true,
-                    allowDeleting: true,
-                    useIcons
-                }
-            });
-
-            this.clock.tick();
-
-            const $buttons = columnsWrapper.getCommandButtons();
-            const $commandCell = $(dataGrid.getCellElement(0, 0));
-
-            // assert
-            assert.ok($commandCell.length, 'command cell is rendered');
-            assert.equal($commandCell.css('white-space'), 'nowrap', 'white-space style');
-            assert.equal($buttons.length, 2, 'command buttons are rendered');
-            $buttons.each((_, button) => {
-                assert.equal($(button).css('display'), 'inline-block', 'display style');
-                assert.equal($(button).css('direction'), 'rtl', 'direction style');
-            });
-        });
-
-        QUnit.test(`Edit command column should not wrap command buttons when useIcons=${useIcons}`, function(assert) {
-            // arrange
-            const dataGrid = createDataGrid({
-                dataSource: [{}],
-                editing: {
-                    allowUpdating: true,
-                    allowDeleting: true,
-                    useIcons
-                },
-                columns: [
-                    {
-                        type: 'buttons'
-                    }
-                ]
-            });
-
-            this.clock.tick();
-
-            // assert
-            const $commandCell = $(dataGrid.getCellElement(0, 0));
-
-            assert.ok($commandCell.length, 'command cell is rendered');
-            assert.equal($commandCell.css('white-space'), 'nowrap', 'white-space style');
-        });
-    });
-
-    QUnit.test('Undelete command buttons should contains aria-label accessibility attribute if rendered as icon and batch edit mode (T755185)', function(assert) {
-    // arrange
-        const columnsWrapper = dataGridWrapper.columns;
-        const dataGrid = createDataGrid({
-            dataSource: [{ id: 0, c0: 'c0' }],
-            columns: [
-                {
-                    type: 'buttons',
-                    buttons: ['undelete']
-                },
-                'id'
-            ],
-            editing: {
-                mode: 'batch',
-                allowUpdating: true,
-                allowDeleting: true,
-                useIcons: true
-            }
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.deleteRow(0);
-        // assert
-        columnsWrapper.getCommandButtons().each((_, button) => {
-            const ariaLabel = $(button).attr('aria-label');
-            assert.ok(ariaLabel && ariaLabel.length, `aria-label '${ariaLabel}'`);
-        });
-    });
-
     QUnit.test('Customize text called for column only (T653374)', function(assert) {
         createDataGrid({
             columns:
@@ -693,74 +533,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
                 done();
             });
         });
-    });
-
-    QUnit.test('Should not cut border of selected cell by \'Add row\' (T748046)', function(assert) {
-    // arrange
-        const clock = sinon.useFakeTimers();
-        const dataGrid = createDataGrid({
-            width: 400,
-            height: 200,
-            showBorders: true,
-            editing: {
-                mode: 'cell',
-                allowAdding: true
-            },
-            dataSource: [...new Array(20)].map((x, i) => ({ name: i }))
-        });
-
-        clock.tick();
-        const scrollable = $('.dx-scrollable').dxScrollable('instance');
-
-        scrollable.scrollTo({ y: 5 });
-        clock.tick();
-
-        // act
-        dataGrid.addRow();
-        clock.tick();
-
-        // assert
-        if(browser.mozilla) {
-            assert.ok(scrollable.scrollTop() <= 1, 'in mozilla first row is overlayed by parent container');
-        } else {
-            assert.ok(scrollable.scrollTop() <= 0.5, 'first row is not overlayed by parent container');
-        }
-
-        clock.restore();
-    });
-
-    QUnit.test('Added row should be scrolled to the top of the grid (T748046)', function(assert) {
-    // arrange
-        const clock = sinon.useFakeTimers();
-        const dataGrid = createDataGrid({
-            width: 400,
-            height: 200,
-            showBorders: true,
-            editing: {
-                mode: 'cell',
-                allowAdding: true
-            },
-            dataSource: [...new Array(20)].map((x, i) => ({ name: i }))
-        });
-
-        clock.tick();
-        const scrollable = $('.dx-scrollable').dxScrollable('instance');
-
-        scrollable.scrollTo({ y: 20 });
-        clock.tick();
-
-        // act
-        dataGrid.addRow();
-        clock.tick();
-
-        // assert
-        if(browser.mozilla) {
-            assert.ok(scrollable.scrollTop() <= 1, 'in mozilla first row is overlayed by parent container');
-        } else {
-            assert.ok(scrollable.scrollTop() <= 0.5, 'first row is not overlayed by parent container');
-        }
-
-        clock.restore();
     });
 
     QUnit.test('Scroller state', function(assert) {
@@ -1651,188 +1423,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal(visibleRows[expandedRowVisibleIndex + 1].key, 29, 'Check next row key');
     });
 
-    // T315857
-    QUnit.test('Editing should work with classes as data objects', function(assert) {
-    // arrange
-        function DataItem(id, text) {
-            this.id = id;
-            this.text = text;
-        }
-        Object.defineProperty(DataItem.prototype, 'ID', {
-            configurable: true,
-            enumerable: false,
-            get: function() { return this.id; },
-            set: function(value) { this.id = value; }
-        });
-        Object.defineProperty(DataItem.prototype, 'Text', {
-            configurable: true,
-            enumerable: false,
-            get: function() { return this.text; },
-            set: function(value) { this.text = value; }
-        });
-        const dataItem0 = new DataItem(0, 'text0');
-        const dataItem1 = new DataItem(1, 'text1');
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columns: ['ID', 'Text'],
-            dataSource: {
-                store: [ dataItem0, dataItem1 ]
-            },
-            editing: { allowUpdating: true, mode: 'batch' }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.cellValue(1, 1, 'test');
-
-        // assert
-        const rows = dataGrid.getVisibleRows();
-        assert.equal(rows.length, 2);
-        assert.equal(rows[1].data.ID, 1);
-        assert.equal(rows[1].data.Text, 'test');
-        assert.deepEqual(rows[1].values, [1, 'test']);
-    });
-
-    // T613804
-    QUnit.test('Editing should work with classes as data objects contains readonly properties', function(assert) {
-    // arrange
-        function DataItem(id, text) {
-            this.id = id;
-            this.text = text;
-        }
-        Object.defineProperty(DataItem.prototype, 'ID', {
-            configurable: true,
-            enumerable: true,
-            get: function() { return this.id; }
-        });
-        Object.defineProperty(DataItem.prototype, 'Text', {
-            configurable: true,
-            enumerable: false,
-            get: function() { return this.text; },
-            set: function(value) { this.text = value; }
-        });
-        const dataItem0 = new DataItem(0, 'text0');
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columns: ['ID', 'Text'],
-            dataSource: {
-                store: [ dataItem0 ]
-            },
-            editing: { allowUpdating: true, mode: 'batch' }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.cellValue(0, 1, 'test');
-
-        // assert
-        const rows = dataGrid.getVisibleRows();
-        assert.equal(rows.length, 1);
-        assert.equal(rows[0].data.ID, 0);
-        assert.equal(rows[0].data.Text, 'test');
-        assert.deepEqual(rows[0].values, [0, 'test']);
-    });
-
-    // T643455
-    QUnit.test('Editing should works with nested readonly property', function(assert) {
-    // arrange
-        function ItemConfig() {
-            this._enable = false;
-        }
-
-        Object.defineProperty(ItemConfig.prototype, 'enable', {
-            get: function() {
-                return this._enable;
-            },
-            set: function(value) {
-                this._enable = value;
-            }
-        });
-
-        function Item(name) {
-            this.name = name;
-            this._config = new ItemConfig();
-        }
-
-        Object.defineProperty(Item.prototype, 'config', {
-            get: function() {
-                return this._config;
-            }
-        });
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columns: ['config.enable', 'name'],
-            dataSource: {
-                store: [ new Item('Test') ]
-            },
-            editing: { allowUpdating: true, mode: 'batch' }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.cellValue(0, 0, true);
-
-        // assert
-        const rows = dataGrid.getVisibleRows();
-        assert.equal(rows.length, 1);
-        assert.deepEqual(rows[0].data.config.enable, true, 'nested property is assigned');
-        assert.ok(rows[0].data.config instanceof ItemConfig, 'config type');
-        assert.deepEqual(rows[0].values, [true, 'Test']);
-    });
-
-    // T613804
-    QUnit.test('calculateCellValue for edited cell fires twice and at the second time contains full data row as an argument', function(assert) {
-    // arrange
-        function DataItem(id, text) {
-            this.id = id;
-            this.text = text;
-        }
-        Object.defineProperty(DataItem.prototype, 'ID', {
-            configurable: true,
-            enumerable: true,
-            get: function() { return this.id; }
-        });
-        Object.defineProperty(DataItem.prototype, 'Text', {
-            configurable: true,
-            enumerable: false,
-            get: function() { return this.text; },
-            set: function(value) { this.text = value; }
-        });
-        const dataItem0 = new DataItem(0, 'text0');
-        let counter = 0;
-        let modifiedData;
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            columns: [
-                'ID',
-                {
-                    dataField: 'Text',
-                    calculateCellValue: function(data) {
-                        if(data.Text === 'test') {
-                            ++counter;
-                            modifiedData = data;
-                        }
-                    }
-                }
-            ],
-            dataSource: {
-                store: [ dataItem0 ]
-            },
-            editing: { allowUpdating: true, mode: 'batch' }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.cellValue(0, 1, 'test');
-        dataGrid.closeEditCell();
-
-        this.clock.tick();
-
-        // assert
-        assert.equal(counter, 2);
-        assert.equal(dataItem0.Text, 'text0');
-        assert.equal(modifiedData.ID, 0);
-        assert.equal(modifiedData.Text, 'test');
-        assert.ok(modifiedData instanceof DataItem, 'modifiedData is instance of DataItem');
-    });
-
     QUnit.test('Group panel should set correct \'max-width\' after clear grouping', function(assert) {
         const dataGrid = $('#dataGrid').dxDataGrid({
             dataSource: {
@@ -2072,55 +1662,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal($element.find('.field1-header').length, 1, 'headerCellTemplate attached to grid');
     });
 
-    // T410328
-    QUnit.test('Edit cell by click when grid is created in dxForm', function(assert) {
-    // arrange
-        let dataGrid;
-
-        $('#form').dxForm({
-            items: [{
-                template: function(options, container) {
-                    dataGrid = $('<div>').appendTo(container).dxDataGrid({
-                        loadingTimeout: undefined,
-                        dataSource: [{ firstName: 1, lastName: 2 }],
-                        editing: {
-                            allowUpdating: true,
-                            mode: 'cell'
-                        }
-                    }).dxDataGrid('instance');
-                }
-            }]
-        });
-
-        // act
-        $(dataGrid.$element().find('.dx-data-row > td').eq(0)).trigger('dxclick');
-        this.clock.tick();
-
-        // assert
-        assert.equal($(dataGrid.$element()).find(TEXTEDITOR_INPUT_SELECTOR).length, 1, 'one editor is shown');
-    });
-
-    QUnit.test('Edit cell by click if repaintChangesOnly is enabled', function(assert) {
-    // arrange
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: [{ firstName: 1, lastName: 2 }],
-            loadingTimeout: undefined,
-            repaintChangesOnly: true,
-            editing: {
-                allowUpdating: true,
-                mode: 'cell'
-            }
-        });
-
-        const $cell = $dataGrid.find('.dx-data-row > td').eq(0);
-
-        // act
-        pointerMock($cell).start().down().up();
-
-        // assert
-        assert.equal($dataGrid.find(TEXTEDITOR_INPUT_SELECTOR).length, 1, 'one editor is shown');
-    });
-
     QUnit.test('export.enabled: true, allowExportSelectedData: true -> check export menu icons (T757579)', function(assert) {
         $('#dataGrid').dxDataGrid({
             export: {
@@ -2189,105 +1730,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
             assert.notOk($loadPanel.hasClass('dx-state-invisible'), 'load panel is visible');
         });
-    });
-
-    QUnit.test('Add row to empty dataGrid - freeSpaceRow element is hidden', function(assert) {
-    // arrange
-        const $grid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [],
-            editing: {
-                allowAdding: true,
-                allowUpdating: true,
-                allowDeleting: true,
-                mode: 'row'
-            },
-            columns: [{ dataField: 'firstName', width: 100 }, { dataField: 'lastName', width: 100 }, { dataField: 'room', width: 100 }, { dataField: 'birthDay', width: 100 }]
-        });
-        const gridInstance = $grid.dxDataGrid('instance');
-
-        // act
-        gridInstance.insertRow();
-        this.clock.tick();
-
-        // assert
-        const $freeSpaceRow = $grid.find('.dx-freespace-row');
-        const $noDataElement = $grid.find('.dx-datagrid-nodata');
-
-        assert.ok(!$freeSpaceRow.is(':visible'), 'Free space row is hidden');
-        assert.ok(!$noDataElement.is(':visible'), 'No data element is hidden');
-    });
-
-    // T744592
-    QUnit.test('freeSpaceRow height should not be changed after editing next cell', function(assert) {
-    // arrange
-        const $grid = $('#dataGrid').dxDataGrid({
-            dataSource: [
-                { id: 1, field1: 'field1' },
-                { id: 2, field1: 'field1' },
-                { id: 3, field1: 'field1' }
-            ],
-            paging: {
-                pageSize: 2
-            },
-            keyExpr: 'id',
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            }
-        });
-        const dataGrid = $grid.dxDataGrid('instance');
-
-        this.clock.tick();
-
-        dataGrid.pageIndex(1);
-        this.clock.tick();
-        dataGrid.cellValue(0, 'field1', 'updated');
-        this.clock.tick();
-        dataGrid.saveEditData();
-
-        // act
-        dataGrid.focus(dataGrid.getCellElement(0, 'field1'));
-
-        // assert
-        assert.ok($grid.find('.dx-freespace-row').is(':visible'), 'Free space row is visible');
-        assert.equal(dataGrid.totalCount(), -1, 'totalCount');
-        assert.equal(dataGrid.getController('data').isLoading(), true, 'isLoading');
-    });
-
-    // T751778
-    QUnit.test('row should not dissapear after insert if dataSource was assigned during saving', function(assert) {
-    // arrange
-        const array = [{ id: '1' }];
-        const $grid = $('#dataGrid').dxDataGrid({
-            dataSource: array,
-            editing: {
-                mode: 'cell',
-                allowAdding: true
-            },
-            keyExpr: 'id',
-            loadingTimeout: 100
-        });
-        const dataGrid = $grid.dxDataGrid('instance');
-
-        // act
-        this.clock.tick(100);
-
-        dataGrid.addRow();
-        dataGrid.cellValue(0, 0, '2');
-        this.clock.tick();
-        dataGrid.closeEditCell();
-        this.clock.tick();
-        dataGrid.option('dataSource', array);
-        this.clock.tick();
-
-        // assert
-        assert.equal($(dataGrid.getCellElement(0, 0)).find('.dx-texteditor-input').val(), '2', 'first row doesn\'t dissapear');
-        assert.equal($(dataGrid.getCellElement(1, 0)).text(), '1', 'second row cell text');
-        // act
-        this.clock.tick(100);
-        assert.equal($(dataGrid.getCellElement(0, 0)).text(), '1', 'first row doesn\'t dissapear');
-        assert.equal($(dataGrid.getCellElement(1, 0)).text(), '2', 'second row cell text');
     });
 
     // T615174
@@ -2821,29 +2263,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal(cellClickArgs[0].key, 1, 'clicked row key');
     });
 
-    QUnit.test('Edit row with the underscore template when the editForm mode is enabled', function(assert) {
-    // arrange
-        const data = [{ firstName: 'Super', lastName: 'Man' }, { firstName: 'Super', lastName: 'Zi' }];
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'form',
-                allowUpdating: true
-            },
-            columnHidingEnabled: true,
-            dataSource: data,
-            columns: ['firstName', { dataField: 'lastName', editCellTemplate: $('#scriptTestTemplate1') }]
-        });
-        const instance = $dataGrid.dxDataGrid('instance');
-
-        // act
-        instance.editRow(0);
-        this.clock.tick();
-
-        // assert
-        assert.equal($dataGrid.find('.dx-form #template1').text(), 'Template1', 'the underscore template is rendered correctly');
-    });
-
     QUnit.test('DataGrid - A fixed rows should be synchronized after change column width if wordWrapEnabled and height are set (T830739)', function(assert) {
     // arrange
         const rowsViewWrapper = dataGridWrapper.rowsView;
@@ -3131,95 +2550,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.strictEqual(dataGrid.getView('rowsView').getScrollbarWidth(true), 0);
     });
 
-    // T386755
-    QUnit.test('column headers visibility when hide removing row in batch editing mode', function(assert) {
-    // arrange, act
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            width: 1000,
-            dataSource: [{ col1: '1', col2: '2' }],
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'batch',
-                allowDeleting: true
-            },
-            onCellPrepared: function(e) {
-                assert.equal(typeUtils.isRenderer(e.cellElement), !!config().useJQuery, 'cellElement is correct');
-                if(e.rowType === 'data' && e.column.command === 'edit' && e.row.removed) {
-                    $(e.cellElement).parent().css({ display: 'none' });
-                }
-            }
-        });
-        const dataGrid = $dataGrid.dxDataGrid('instance');
-
-        dataGrid.deleteRow(0);
-
-        // assert
-        assert.strictEqual(dataGrid.getView('rowsView').getScrollbarWidth(), 0, 'vertical scrollbar width');
-        assert.strictEqual($dataGrid.find('.dx-datagrid-headers').css('paddingRight'), '0px', 'no headers right padding');
-    });
-
-    // T852171
-    QUnit.test('numberbox input right and left paddings should be equal if spin buttons are showed', function(assert) {
-    // arrange
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: [{ field: 30 }],
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: [{
-                dataField: 'field',
-                dataType: 'number',
-                editorOptions: {
-                    showSpinButtons: true
-                }
-            }]
-        });
-
-        const dataGrid = $dataGrid.dxDataGrid('instance');
-
-        // act
-        dataGrid.editCell(0, 0);
-
-        const $input = $($dataGrid.find('.dx-editor-cell').find('.dx-texteditor-input'));
-
-        // assert
-        assert.equal($input.length, 1, 'input');
-        assert.equal($input.css('padding-right'), $input.css('padding-left'), 'paddings are equal');
-    });
-
-    // T712073
-    QUnit.test('Delete two added rows after selection', function(assert) {
-    // arrange, act
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            width: 1000,
-            dataSource: [{ id: 1 }],
-            keyExpr: 'id',
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'batch',
-                allowAdding: true,
-                allowDeleting: true
-            }
-        });
-        const dataGrid = $dataGrid.dxDataGrid('instance');
-
-        // act
-        dataGrid.addRow();
-        dataGrid.addRow();
-        dataGrid.selectRows(1);
-
-        $dataGrid.find('.dx-link-delete').first().trigger('click');
-        this.clock.tick();
-        $dataGrid.find('.dx-link-delete').first().trigger('click');
-        this.clock.tick();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 1, 'row count');
-        assert.strictEqual($dataGrid.find('.dx-data-row').length, 1, 'visible data row count');
-    });
-
     QUnit.test('Disable rows hover', function(assert) {
     // arrange
         const $dataGrid = $('#dataGrid').dxDataGrid({
@@ -3356,38 +2686,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
                 assert.equal(dataGrid.pageIndex(), 1, 'Page index');
             });
         });
-    });
-
-    // T803784
-    QUnit.test('Command cell should not have dx-hidden-cell class if it is not fixed', function(assert) {
-    // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ field: 'data' }],
-            columns: [{
-                dataField: 'field',
-                caption: 'fixed',
-                fixed: true
-            }, {
-                dataField: 'field',
-                caption: 'not fixed'
-            }, {
-                type: 'buttons',
-                fixed: false,
-                buttons: ['edit']
-            }],
-            editing: {
-                mode: 'row',
-                allowUpdating: true,
-                useIcons: true
-            }
-        }).dxDataGrid('instance');
-
-        // assert
-        const rows = dataGrid.getRowElement(0);
-
-        assert.equal(Math.floor($(rows[0]).find('td').eq(0).width()), Math.floor($(rows[1]).find('td').eq(0).width()), 'widths are equal');
-        assert.notOk($('.dx-command-edit').eq(1).hasClass('dx-hidden-cell'), 'cell does not have class dx-hidden-cell');
     });
 
     QUnit.test('The navigateToRow method should not affect horizontal scrolling', function(assert) {
@@ -4350,154 +3648,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.ok(dataGrid.getScrollable().scrollTop() > 0, 'scrollTop is not reseted');
     });
 
-    [false, true].forEach(function(grouping) {
-        QUnit.test(`loading data on scroll after deleting several rows if scrolling mode is infinite and refreshMode is repaint ${grouping ? 'and if grouping ' : ''}(T862268)`, function(assert) {
-            // arrange
-            const array = [];
-
-            for(let i = 1; i <= 50; i++) {
-                array.push({ id: i, group: Math.floor(i / 10) });
-            }
-
-            const dataGrid = $('#dataGrid').dxDataGrid({
-                height: 100,
-                dataSource: array,
-                keyExpr: 'id',
-                editing: {
-                    allowDeleting: true,
-                    texts: {
-                        confirmDeleteMessage: ''
-                    },
-                    refreshMode: 'repaint'
-                },
-                scrolling: {
-                    mode: 'infinite',
-                    useNative: false
-                },
-                columns: ['id', {
-                    dataField: 'group',
-                    groupIndex: grouping ? 0 : undefined }
-                ],
-                loadingTimeout: undefined
-            }).dxDataGrid('instance');
-
-            const firstDataRowIndex = grouping ? 1 : 0;
-            // act
-            dataGrid.deleteRow(firstDataRowIndex);
-            dataGrid.deleteRow(firstDataRowIndex);
-            dataGrid.getScrollable().scrollTo({ y: 10000 });
-
-            // assert
-            let rows = dataGrid.getVisibleRows();
-            assert.equal(dataGrid.totalCount(), 38, 'totalCount');
-            assert.equal(rows.length, 38, 'visible row count');
-            assert.equal(rows[firstDataRowIndex].key, 3, 'row 0');
-            assert.equal(rows[18].key, grouping ? 19 : 21, 'row 18');
-            assert.equal(rows[37].key, grouping ? 36 : 40, 'row 37');
-
-            // act
-            dataGrid.refresh();
-
-            // assert
-            rows = dataGrid.getVisibleRows();
-            assert.equal(dataGrid.totalCount(), 20, 'totalCount');
-            assert.equal(rows.length, 20, 'visible row count');
-            assert.equal(rows[firstDataRowIndex].key, 3, 'row 0');
-        });
-    });
-
-    QUnit.test('loading data on scroll after deleting several rows if scrolling mode is infinite, rowRenderingMode is virtual and refreshMode is repaint (T862268)', function(assert) {
-        // arrange
-        const array = [];
-
-        for(let i = 1; i <= 150; i++) {
-            array.push({ id: i });
-        }
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            height: 100,
-            dataSource: array,
-            keyExpr: 'id',
-            editing: {
-                allowDeleting: true,
-                texts: {
-                    confirmDeleteMessage: ''
-                },
-                refreshMode: 'repaint'
-            },
-            paging: {
-                pageSize: 50
-            },
-            scrolling: {
-                mode: 'infinite',
-                rowRenderingMode: 'virtual',
-                useNative: false
-            },
-            columns: ['id'],
-            loadingTimeout: undefined
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.getScrollable().scrollTo({ y: 10000 });
-        dataGrid.getScrollable().scrollTo({ y: 0 });
-        dataGrid.deleteRow(0);
-        dataGrid.deleteRow(0);
-        dataGrid.deleteRow(0);
-        dataGrid.getScrollable().scrollTo({ y: 10000 });
-        dataGrid.getScrollable().scrollTo({ y: 10000 });
-
-        // assert
-        const rows = dataGrid.getVisibleRows();
-        assert.equal(dataGrid.totalCount(), 147, 'totalCount');
-        assert.equal(rows[rows.length - 1].key, 150, 'last row key');
-    });
-
-
-    ['repaint', 'reshape'].forEach((refreshMode) => {
-        QUnit.test(`loading data on scroll after adding row if scrolling mode is infinite and refreshMode is ${refreshMode} (T914296)`, function(assert) {
-            // arrange
-            const array = [];
-
-            for(let i = 1; i <= 150; i++) {
-                array.push({ id: i });
-            }
-
-            const dataGrid = $('#dataGrid').dxDataGrid({
-                height: 400,
-                dataSource: array,
-                keyExpr: 'id',
-                editing: {
-                    mode: 'cell',
-                    allowAdding: true,
-                    refreshMode: refreshMode
-                },
-                paging: {
-                    pageSize: 50
-                },
-                scrolling: {
-                    mode: 'infinite',
-                    useNative: false
-                },
-                columns: ['id'],
-                loadingTimeout: undefined
-            }).dxDataGrid('instance');
-
-            // act
-            dataGrid.getScrollable().scrollTo({ y: 10000 });
-            dataGrid.getScrollable().scrollTo({ y: 0 });
-            dataGrid.addRow();
-            dataGrid.saveEditData();
-            dataGrid.getScrollable().scrollTo({ y: 10000 });
-            dataGrid.getScrollable().scrollTo({ y: 10000 });
-            dataGrid.getScrollable().scrollTo({ y: 10000 });
-
-            // assert
-            const rows = dataGrid.getVisibleRows();
-            assert.strictEqual(dataGrid.totalCount(), refreshMode === 'repaint' ? 152 : 151, 'totalCount'); // TODO: Fix duplicate added row when editing.refreshMode = 'repaint'
-            assert.strictEqual(rows[rows.length - 2].key, 150, 'penultimate row key');
-        });
-    });
-
     QUnit.test('height from extern styles', function(assert) {
     // arrange, act
         const $dataGrid = $('#dataGrid').addClass('fixed-height').dxDataGrid({
@@ -4884,195 +4034,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal($dataGridTables.length, 2);
         assert.equal($dataGridTables.eq(0).find('.dx-row').first().find('td').last()[0].getBoundingClientRect().width, 700);
         assert.equal($dataGridTables.eq(1).find('.dx-row').first().find('td').last()[0].getBoundingClientRect().width, 700);
-    });
-
-    // T422575, T411642
-    QUnit.test('column widths should be synchronized when scrolling mode is virtual and lookup column and edit column are exist', function(assert) {
-    // arrange, act
-        let contentReadyCallCount = 0;
-        const $dataGrid = $('#dataGrid').dxDataGrid({
-            onContentReady: function(e) {
-                contentReadyCallCount++;
-            },
-            width: 1000,
-            height: 200,
-            dataSource: [{ field1: 1, field2: 2 }, { field1: 3, field2: 4 }, { field1: 5, field2: 6 }, { field1: 7, field2: 8 }],
-            scrolling: { mode: 'virtual' },
-            paging: {
-                pageSize: 3
-            },
-            editing: {
-                allowUpdating: true
-            },
-            columns: [
-                {
-                    dataField: 'field1', lookup:
-                    {
-                        displayExpr: 'text',
-                        valueExpr: 'value',
-                        dataSource: {
-                            load: function() {
-                                const d = $.Deferred();
-
-                                setTimeout(function() {
-                                    d.resolve([{ value: 1, text: 'text 1' }, { value: 2, text: 'text 2' }]);
-                                });
-
-                                return d;
-                            }
-                        }
-                    }
-                },
-                { dataField: 'field2' }
-            ]
-        });
-
-        this.clock.tick();
-
-        const $dataGridTables = $dataGrid.find('.dx-datagrid-table');
-        // assert
-        assert.equal(contentReadyCallCount, 1);
-        assert.equal($dataGridTables.length, 2);
-        assert.equal($dataGridTables.eq(0).find('.dx-row').first().find('td')[0].getBoundingClientRect().width, $dataGridTables.eq(1).find('.dx-row').first().find('td')[0].getBoundingClientRect().width);
-
-        assert.equal($dataGridTables.eq(0).find('.dx-row').first().find('td')[1].getBoundingClientRect().width, $dataGridTables.eq(1).find('.dx-row').first().find('td')[1].getBoundingClientRect().width);
-    });
-
-    // T833061
-    QUnit.test('Last row should be correct after editing other row\'s cell if scrolling and rendering are virtual', function(assert) {
-    // arrange
-        const dataSource = [];
-
-        for(let i = 0; i < 40; i++) {
-            dataSource.push({ field: i });
-        }
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource,
-            height: 150,
-            editing: {
-                enabled: true,
-                mode: 'cell',
-                allowUpdating: true
-            },
-            scrolling: {
-                rowRenderingMode: 'virtual',
-                mode: 'virtual',
-                useNative: false
-            }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.getScrollable().scrollTo({ y: 1500 });
-
-        dataGrid.editCell(8, 0);
-
-        const visibleRows = dataGrid.getVisibleRows();
-
-        // assert
-        assert.notOk(visibleRows[-1], 'no visible row with index -1');
-        assert.equal($(dataGrid.getCellElement(9, 0)).text(), '39', 'last row is correct');
-    });
-
-    // T833061
-    QUnit.test('Edit cell after editing another cell and scrolling down should work correctly if scrolling and rendering are virtual', function(assert) {
-    // arrange
-        const dataSource = [];
-        let $editedRow;
-        let $input;
-
-        for(let i = 0; i < 100; i++) {
-            dataSource.push({ field: i });
-        }
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource,
-            height: 440,
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            scrolling: {
-                rowRenderingMode: 'virtual',
-                mode: 'virtual',
-                useNative: false
-            }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.editCell(8, 0);
-
-        dataGrid.getScrollable().scrollTo({ y: 1000 });
-
-        dataGrid.editCell(5, 0);
-
-        const visibleRows = dataGrid.getVisibleRows();
-
-        const hasNegativeIndexes = Object.keys(visibleRows).some(rowIndex => rowIndex < 0);
-
-        const $rows = dataGrid.$element().find('.dx-data-row');
-
-        // assert
-        assert.notOk(hasNegativeIndexes, 'no visible rows with index < 0');
-
-        const startValue = parseInt($rows.eq(0).text());
-
-        assert.equal(startValue, 25, 'visible row #1 is correct');
-
-        for(let i = 1; i < $rows.length; i++) {
-            if(i !== 5) {
-                assert.equal(parseInt($rows.eq(i).text()), startValue + i, `visible row's #${i + 1} text`);
-            } else {
-                $editedRow = $rows.eq(i);
-                $input = $editedRow.find('input');
-
-                assert.ok($editedRow.find('.dx-editor-cell').length, 'row has editor');
-                assert.equal(parseInt($input.val()), startValue + i, `visible row's #${i + 1} input value`);
-            }
-        }
-    });
-
-    // T833071
-    QUnit.test('Click on cell should open editor after scrolling grid down if scrolling and rendering are virtual and repaintChangesOnly is true', function(assert) {
-    // arrange
-        const dataSource = [];
-
-        for(let i = 0; i < 100; i++) {
-            dataSource.push({ field: i });
-        }
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource,
-            height: 150,
-            editing: {
-                enabled: true,
-                mode: 'cell',
-                allowUpdating: true
-            },
-            repaintChangesOnly: true,
-            scrolling: {
-                rowRenderingMode: 'virtual',
-                mode: 'virtual',
-                useNative: false
-            }
-        }).dxDataGrid('instance');
-
-        // act
-        dataGrid.getScrollable().scrollTo({ y: 3000 });
-
-        dataGrid.editCell(1, 0);
-
-        const visibleRows = dataGrid.getVisibleRows();
-        const $rows = dataGrid.$element().find('.dx-data-row');
-        const $editorCell = $rows.eq(1).find('.dx-editor-cell');
-
-        // assert
-        assert.ok($editorCell.length, 'row has editor');
-        assert.equal($editorCell.find('input').val(), '86', 'input value');
-        assert.notOk(visibleRows[-1], 'no visible row with index -1');
     });
 
     // T352218
@@ -6456,34 +5417,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.equal(cellTemplateCallCount, 2, 'cellTemplate call count');
     });
 
-    // T719938
-    QUnit.test('No error after adding row and virtual scrolling', function(assert) {
-    // act
-        const dataGrid = createDataGrid({
-            height: 50,
-            paging: { pageSize: 2 },
-            scrolling: { mode: 'virtual' },
-            columns: ['id'],
-            keyExpr: 'id',
-            dataSource: [...Array(10)].map((_, i) => { return { id: i + 1 }; })
-        });
-
-        this.clock.tick();
-        dataGrid.addRow();
-        this.clock.tick();
-        dataGrid.pageIndex(1);
-        this.clock.tick();
-        dataGrid.pageIndex(2);
-        this.clock.tick();
-        dataGrid.pageIndex(3);
-        this.clock.tick();
-        dataGrid.pageIndex(0);
-        this.clock.tick();
-
-        // assert
-        assert.strictEqual($($(dataGrid.$element()).find('.dx-error-row')).length, 0, 'no errors');
-    });
-
     QUnit.test('Not raise error if key field is null', function(assert) {
     // act
         const dataGrid = createDataGrid({
@@ -6711,110 +5644,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         // assert
         assert.equal(toolbarItemOffset, $(dataGrid.$element()).find('.dx-datagrid-search-panel').offset().top, 'toolbar sarch panel is aligned');
         assert.equal(toolbarItemOffset, $(dataGrid.$element()).find('.dx-toolbar .dx-datebox').offset().top, 'toolbar custom item is aligned');
-    });
-
-    // T809423
-    QUnit.test('Toolbar should not be rerendered if editing.popup options were changed', function(assert) {
-        const onToolbarPreparingSpy = sinon.spy();
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [],
-            onToolbarPreparing: onToolbarPreparingSpy,
-            editing: {
-                mode: 'popup'
-            }
-        });
-
-        dataGrid.option('editing.popup', {});
-
-        assert.equal(onToolbarPreparingSpy.callCount, 1, 'onToolbarPreparing call count');
-    });
-
-    // T558301
-    QUnit.testInActiveWindow('Height virtual table should be updated to show validation message when there is a single row and virtual scrolling is enabled', function(assert) {
-    // arrange
-        let $tableElements;
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ Test: '' }],
-            editing: {
-                mode: 'batch',
-                allowUpdating: true
-            },
-            scrolling: {
-                mode: 'virtual'
-            },
-            columns: [{
-                dataField: 'Test',
-                validationRules: [{ type: 'required' }]
-            }]
-        });
-
-        // assert
-        $tableElements = dataGrid.$element().find('.dx-datagrid-rowsview').find('table');
-        assert.roughEqual($tableElements.eq(0).outerHeight(), 35, 3, 'height main table');
-
-        // act
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-
-        // assert
-        $tableElements = dataGrid.$element().find('.dx-datagrid-rowsview').find('table');
-        assert.roughEqual($tableElements.eq(0).outerHeight(), 68, 3.01, 'height main table');
-
-        dataGrid.closeEditCell();
-        this.clock.tick();
-
-        // assert
-        $tableElements = dataGrid.$element().find('.dx-datagrid-rowsview').find('table');
-        assert.roughEqual($tableElements.eq(0).outerHeight(), 35, 3, 'height main table');
-    });
-
-    QUnit.test('Error row is not hidden when rowKey is undefined by mode is cell', function(assert) {
-    // arrange
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{
-                'ID': 1,
-                'FirstName': 'John',
-                'LastName': 'Heart',
-                'Prefix': 'Mr.',
-                'Position': 'CEO',
-                'BirthDate': '1964/03/16',
-                'HireDate': '1995/01/15',
-                'Address': '351 S Hill St.',
-                'StateID': 5
-            }],
-            keyExpr: 'myFakeKey',
-            paging: {
-                enabled: false
-            },
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: ['Prefix', 'FirstName']
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-
-        $('input')
-            .val('new')
-            .change();
-
-        this.clock.tick();
-
-        dataGrid.editCell(0, 1);
-        this.clock.tick();
-
-        // assert
-        assert.equal($('.dx-error-message').length, 1, 'Error message is shown');
     });
 
     // T689367
@@ -7301,77 +6130,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
         assert.strictEqual($(dataGrid.$element()).find('.dx-error-row').length, 0, 'error row is hidden');
     });
 
-    // T807774
-    QUnit.test('Editor should be rendered for hidden columns while editing in row mode with repaintChangesOnly', function(assert) {
-    // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'row',
-                allowUpdating: true
-            },
-            repaintChangesOnly: true,
-            columnHidingEnabled: true,
-            width: 200,
-            dataSource: [{ field1: '1', field2: '2' }]
-        }).dxDataGrid('instance');
-
-        // act
-        $('.dx-datagrid .dx-datagrid-adaptive-more')
-            .eq(0)
-            .trigger('dxclick');
-
-        $(dataGrid.getRowElement(0)).find('.dx-command-edit > .dx-link-edit').trigger(pointerEvents.up).click();
-        this.clock.tick();
-
-        const $firstRowEditors = $(dataGrid.getRowElement(1)).find('.dx-texteditor');
-
-        // assert
-        assert.ok($firstRowEditors.length, 'row has editor');
-        assert.notOk($firstRowEditors.eq(0).parent().hasClass('dx-adaptive-item-text'), 'editor\'s parent does not have class');
-
-        // act
-        $(dataGrid.getRowElement(0)).find('.dx-command-edit > .dx-link-cancel').trigger(pointerEvents.up).click();
-        this.clock.tick();
-
-        // assert
-        assert.notOk($(dataGrid.getRowElement(1)).find('.dx-texteditor').length, 'row doesn\'t have editor');
-    });
-
-    // T865715
-    QUnit.test('Row\'s height should be correct after updateDimensions while editing with popup edit mode', function(assert) {
-        // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            width: 300,
-            columnHidingEnabled: true,
-            keyExpr: 'ID',
-            wordWrapEnabled: true,
-            editing: {
-                allowUpdating: true,
-                mode: 'popup'
-            },
-            dataSource: [{
-                ID: 1,
-                Comment: 'very long text very long text very long text very long text very long text very long text very long text very long text'
-            }, {
-                ID: 2,
-                Comment: 'very long text very long text very long text very long text very long text very long text very long text very long text'
-            }]
-        }).dxDataGrid('instance');
-
-        // act
-        const $firstRow = $(dataGrid.getRowElement(0));
-        const firstRowHeight = $firstRow.height();
-
-        dataGrid.editRow(0);
-        dataGrid.updateDimensions();
-
-        // assert
-        assert.equal($firstRow.height(), firstRowHeight, 'first row\'s height');
-        assert.ok($firstRow.find('td').eq(1).hasClass('dx-datagrid-hidden-column'), 'column hiding class');
-    });
-
     QUnit.test('Scrollable should have the correct padding when the grid inside the ScrollView', function(assert) {
     // arrange, act
         $('#container').dxScrollView({
@@ -7608,46 +6366,6 @@ QUnit.module('Initialization', baseModuleConfig, () => {
 
 
 QUnit.module('Virtual row rendering', baseModuleConfig, () => {
-
-    QUnit.test('editing should starts correctly if scrolling mode is virtual', function(assert) {
-    // arrange, act
-        const array = [];
-
-        for(let i = 1; i <= 50; i++) {
-            array.push({ id: i });
-        }
-
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            height: 100,
-            dataSource: array,
-            keyExpr: 'id',
-            onRowPrepared: function(e) {
-                $(e.rowElement).css('height', 50);
-            },
-            editing: {
-                mode: 'row',
-                allowUpdating: true
-            },
-            scrolling: {
-                mode: 'virtual',
-                rowRenderingMode: 'virtual',
-                useNative: false
-            }
-        }).dxDataGrid('instance');
-
-        this.clock.tick();
-
-        // act
-        dataGrid.getScrollable().scrollTo({ top: 500 });
-        dataGrid.editRow(1);
-
-        // assert
-        const visibleRows = dataGrid.getVisibleRows();
-        assert.equal(visibleRows.length, 15, 'visible row count');
-        assert.equal(visibleRows[0].key, 6, 'first visible row key');
-        assert.equal($(dataGrid.getRowElement(1, 0)).find('.dx-texteditor').length, 1, 'row has editor');
-    });
-
     // T644981
     QUnit.test('grouping should works correctly if row rendering mode is virtual and dataSource is remote', function(assert) {
     // arrange, act
@@ -8400,29 +7118,6 @@ QUnit.module('Rendered on server', baseModuleConfig, () => {
         // assert
         assert.equal(dataGrid.getVisibleRows().length, 1, '1 visible row');
         assert.equal(dataGrid.$element().find('.dx-virtual-row').length, 0, 'no virtual rows');
-    });
-
-    // T621703
-    QUnit.testInActiveWindow('Edit cell on onContentReady', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [{ firstName: 'Andrey', lastName: 'Prohorov' }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            onContentReady: function(e) {
-            // act
-                e.component.editCell(0, 1);
-            }
-        });
-
-        this.clock.tick();
-
-        // assert
-        const $cellElement = $(dataGrid.getCellElement(0, 1));
-        assert.ok($cellElement.hasClass('dx-editor-cell'), 'cell has editor');
-        assert.ok($cellElement.find('.dx-texteditor-input').is(':focus'), 'cell editor is focused');
     });
 });
 
@@ -10399,27 +9094,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
         assert.strictEqual(visibleColumns[1].groupIndex, 1, 'groupIndex of the second column');
     });
 
-    // T582855
-    QUnit.test('change editing.allowAdding with onCellPrepared and dataSource options should update add row button', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({});
-
-        // act
-        dataGrid.option({
-            editing: {
-                allowAdding: true
-            },
-            onCellPrepared: function() {},
-            dataSource: []
-        });
-
-        this.clock.tick();
-
-        // assert
-        const $addRowButton = dataGrid.$element().find('.dx-datagrid-addrow-button');
-        assert.strictEqual($addRowButton.length, 1, 'add row button is rendered');
-    });
-
     // T689294
     QUnit.test('onContentReady when there is no dataSource and stateStoring is enabled', function(assert) {
     // arrange
@@ -10442,94 +9116,6 @@ QUnit.module('Assign options', baseModuleConfig, () => {
 
         // assert
         assert.equal(contentReadyCallCount, 1);
-    });
-
-    // T749733
-    QUnit.test('Change editing.popup option should not reload data', function(assert) {
-    // arrange
-        const lookupLoadingSpy = sinon.spy();
-        const dataGrid = createDataGrid({
-            onInitNewRow: function(e) {
-                e.component.option('editing.popup.title', 'New title');
-            },
-            dataSource: [],
-            editing: {
-                mode: 'popup',
-                allowAdding: true,
-                popup: {
-                    showTitle: true
-                }
-            },
-            columns: [{
-                dataField: 'Task_Assigned_Employee_ID',
-                lookup: {
-                    dataSource: {
-                        load: function() {
-                            lookupLoadingSpy();
-                            const d = $.Deferred();
-                            setTimeout(function() {
-                                d.resolve([]);
-                            }, 100);
-                            return d.promise();
-                        }
-                    },
-                    valueExpr: 'Customer_ID',
-                    displayExpr: 'Customer_Name'
-                }
-            }]
-        });
-        this.clock.tick(100);
-
-        // act
-        dataGrid.addRow();
-        this.clock.tick(100);
-
-        // assert
-        assert.equal(lookupLoadingSpy.callCount, 1, 'lookup is loaded once');
-        assert.equal(dataGrid.getController('editing')._editPopup.option('title'), 'New title', 'popup title is updated');
-    });
-
-    QUnit.test('DataGrid should update editor values in Popup Edit Form if its data was reloaded (T815443)', function(assert) {
-    // arrange
-        let loadCallCount = 0;
-        let changeEditorValue;
-        const data = [{ 'name': 'Alex', 'age': 22, 'id': 1 }];
-        const dataGrid = createDataGrid({
-            dataSource: {
-                key: 'id',
-                load: () => {
-                    if(loadCallCount > 0) {
-                        data[0]['name'] = 'foo';
-                    }
-                    loadCallCount++;
-                    return data;
-                }
-            },
-            editing: {
-                mode: 'popup',
-                allowUpdating: true
-            },
-            onEditorPreparing: function(args) {
-                if(args.parentType === 'dataRow' && args.dataField === 'age') {
-                    changeEditorValue = () => {
-                        args.setValue(30);
-                        args.component.getDataSource().reload();
-                    };
-                }
-            },
-        });
-        this.clock.tick();
-
-        // act
-        dataGrid.editRow(0);
-        this.clock.tick();
-
-        changeEditorValue();
-        this.clock.tick();
-
-        // assert
-        const $popupEditorInput = $('.dx-popup-content').find('.dx-texteditor').eq(0).find('input').eq(0);
-        assert.equal($popupEditorInput.val(), 'foo', 'value changed');
     });
 
     // T824018
@@ -10663,232 +9249,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
 
         // assert
         assert.equal(dataGrid.getView('rowsView')._loadPanel.option('message'), 'Loading...');
-    });
-
-    QUnit.test('insert row', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ id: 1111 }]
-        });
-
-        // act
-        dataGrid.addRow();
-
-        // assert
-        assert.equal($('#dataGrid').find(TEXTEDITOR_INPUT_SELECTOR).length, 1);
-        assert.equal($('#dataGrid').find('.dx-datagrid-rowsview').find('tbody > tr').length, 3, 'inserting row + data row + freespace row');
-    });
-
-    // T652111
-    QUnit.test('add row if dataSource is not defined', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            columns: ['id', 'text']
-        });
-
-        // act
-        dataGrid.addRow();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 0, 'no visible rows');
-    });
-
-    // T722161
-    QUnit.test('add row after scrolling if rowRendringMode is virtual', function(assert) {
-        const array = [];
-        for(let i = 1; i <= 20; i++) {
-            array.push({ id: i, text: 'text' + i });
-        }
-        // arrange, act
-        const dataGrid = createDataGrid({
-            height: 200,
-            dataSource: array,
-            keyExpr: 'id',
-            loadingTimeout: undefined,
-            paging: {
-                pageSize: 10
-            },
-            scrolling: {
-                mode: 'virtual',
-                rowRenderingMode: 'virtual',
-                useNative: false
-            },
-            columns: ['id', 'text']
-        });
-
-        // act
-        dataGrid.pageIndex(1);
-        dataGrid.addRow();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows()[0].key, 6, 'first visible row key');
-        assert.ok(dataGrid.getVisibleRows()[5].isNewRow, 'inserted row exists');
-        assert.deepEqual(dataGrid.getVisibleRows()[5].values, [undefined, undefined], 'inserted row values');
-    });
-
-    QUnit.test('add row without return key', function(assert) {
-    // arrange, act
-        const array = [{ id: 1, name: 'Test 1' }];
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'batch'
-            },
-            dataSource: {
-                key: 'id',
-                load: function() {
-                    return array;
-                },
-                insert: function(values) {
-                    array.push(values);
-                }
-            }
-        });
-
-        // act
-        dataGrid.addRow();
-        dataGrid.saveEditData();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 2, 'visible rows');
-        assert.strictEqual(dataGrid.hasEditData(), false, 'no edit data');
-    });
-
-    QUnit.test('Disable editing buttons after insert a row', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ id: 1111 }],
-            editing: {
-                mode: 'batch',
-                allowAdding: true,
-                allowUpdating: true,
-                allowDelete: true
-            }
-        });
-        let $editButtons = $('#dataGrid .dx-edit-button');
-
-        assert.ok(!$editButtons.eq(0).hasClass('dx-state-disabled'), 'Insert button isn\'t disabled');
-        assert.ok($editButtons.eq(1).hasClass('dx-state-disabled'), 'Save button is disabled');
-        assert.ok($editButtons.eq(2).hasClass('dx-state-disabled'), 'Revert button is disabled');
-
-        // act
-        dataGrid.addRow();
-
-        $editButtons = $('#dataGrid .dx-edit-button');
-
-        // assert
-        assert.ok(!$editButtons.eq(0).hasClass('dx-state-disabled'), 'Insert button isn\'t disabled');
-        assert.ok(!$editButtons.eq(1).hasClass('dx-state-disabled'), 'Save button isn\'t disabled');
-        assert.ok(!$editButtons.eq(2).hasClass('dx-state-disabled'), 'Revert button isn\'t disabled');
-    });
-
-    QUnit.test('insert row when master detail autoExpandAll is active', function(assert) {
-    // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ id: 1111 }],
-            masterDetail: {
-                enabled: true,
-                autoExpandAll: true,
-                template: function(container, options) {
-                    $(container).append($('<div>detail</div>'));
-                }
-            }
-        });
-
-        // act
-        dataGrid.addRow();
-        const rows = $('#dataGrid').find('.dx-datagrid-rowsview').find('tbody > tr');
-
-        // assert
-        assert.ok(rows.eq(0).hasClass('dx-row-inserted'), 'First row is inserted row');
-        assert.ok(rows.eq(1).hasClass('dx-row'), 'Second row has dx-row class');
-        assert.ok(!rows.eq(1).hasClass('dx-master-detail-row'), 'Second row is not master-detail-row');
-        assert.ok(rows.eq(2).hasClass('dx-master-detail-row'), 'Third row is master-detail-row');
-    });
-
-    // T636146
-    QUnit.test('onRowInserted should be called if dataSource is reassigned in loadingChanged', function(assert) {
-        const rowInsertedArgs = [];
-        const dataSource = [{ id: 1 }, { id: 2 }];
-        let isLoadingOccurs;
-        const dataGrid = createDataGrid({
-            keyExpr: 'id',
-            dataSource: dataSource,
-            onRowInserted: function(e) {
-                rowInsertedArgs.push(e);
-            }
-        });
-
-        this.clock.tick(0);
-
-        dataGrid.addRow();
-        dataGrid.cellValue(0, 0, 3);
-        dataGrid.getDataSource().on('loadingChanged', function(isLoading) {
-            if(isLoading && !isLoadingOccurs) {
-                dataGrid.option('dataSource', dataGrid.option('dataSource'));
-                isLoadingOccurs = true;
-            }
-        });
-
-        // act
-        dataGrid.saveEditData();
-        this.clock.tick(0);
-
-        // assert
-        assert.equal(isLoadingOccurs, true, 'loadingChanged is occurs');
-        assert.equal(rowInsertedArgs.length, 1, 'rowInserted is called');
-        assert.deepEqual(rowInsertedArgs[0].data, { id: 3 }, 'rowInserted data arg');
-    });
-
-    QUnit.test('The row should be added after the \'addRow\' method was called in the \'onRowInserted\' event (T650889)', function(assert) {
-    // arrange
-        let $inputElement;
-        let needAddRow = true;
-        const dataGrid = createDataGrid({
-            editing: {
-                mode: 'popup',
-                allowAdding: true,
-                allowUpdating: true
-            },
-            keyExpr: 'name',
-            dataSource: [{ name: 'Alex' }],
-            onRowInserted: function(e) {
-                if(needAddRow) {
-                    needAddRow = !needAddRow;
-                    e.component.addRow();
-                }
-            }
-        });
-
-        this.clock.tick();
-
-        fx.off = false;
-
-        // act
-        dataGrid.addRow();
-        $inputElement = $('.dx-popup-content').find('input').first();
-        $inputElement.val('name1').trigger('change');
-        dataGrid.saveEditData();
-
-        this.clock.tick();
-
-        $inputElement = $('.dx-popup-content').find('input').first();
-        $inputElement.val('name2').trigger('change');
-        dataGrid.saveEditData();
-
-        this.clock.tick();
-
-        // assert
-        const visibleRows = dataGrid.getVisibleRows();
-        assert.equal(visibleRows.length, 3, 'rows count');
-        assert.equal(visibleRows[1].data.name, 'name1', 'added cell value');
-        assert.equal(visibleRows[2].data.name, 'name2', 'added cell value');
-
-        fx.off = true;
     });
 
     QUnit.test('LoadPanel show when grid rendering in detail row', function(assert) {
@@ -11028,263 +9388,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         // assert
         assert.equal(getCellTextAlignByButtonNumber(0), 'right', 'Cell template has correct text-align');
         assert.equal(getCellTextAlignByButtonNumber(1), 'right', 'Detail cell has correct text-align');
-    });
-
-    // T755201
-    QUnit.test('Revert button should appear in cell mode when editing column with boolean dataField and saving is canceled', function(assert) {
-    // arrange
-        createDataGrid({
-            dataSource: [{ value: false, id: 1 }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            onRowUpdating: function(e) {
-                const d = $.Deferred();
-                e.cancel = d.promise();
-
-                setTimeout(function() {
-                    d.resolve(true);
-                });
-            },
-            columns: ['id', { dataField: 'value', allowEditing: true }]
-        });
-        this.clock.tick();
-        this.clock.tick(1000);
-
-        // act
-        $('.dx-checkbox').eq(0).trigger('dxclick');
-        this.clock.tick();
-
-        // assert
-        assert.equal($('.dx-checkbox').eq(0).attr('aria-checked'), 'true', 'checkbox is checked');
-        assert.equal($('.dx-revert-button').length, 1, 'reverse button exists');
-
-        // act
-        $('.dx-revert-button').trigger('dxclick');
-
-        // assert
-        assert.equal($('.dx-checkbox').eq(0).attr('aria-checked'), 'false', 'checkbox is unchecked');
-    });
-
-    // T833456
-    QUnit.test('Click to boolean column should works after editing in another column', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [{ name: 'name 1', value: false, id: 1 }],
-            keyExpr: 'id',
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            repaintChangesOnly: true,
-            columns: ['name', 'value']
-        });
-        this.clock.tick();
-
-        // act
-        dataGrid.editCell(0, 0);
-
-        // act
-        const $input = $('.dx-texteditor-input').eq(0);
-        $input.val('test');
-
-        // act
-        let $checkbox = $('.dx-checkbox').eq(0);
-        $input.trigger('change');
-        $checkbox.trigger('dxpointerdown');
-        this.clock.tick();
-        $checkbox.trigger('dxclick');
-        this.clock.tick();
-
-        // assert
-        $checkbox = $('.dx-checkbox').eq(0);
-        assert.equal($checkbox.attr('aria-checked'), 'true', 'checkbox is checked');
-        assert.ok($checkbox.hasClass('dx-state-focused'), 'checkbox is focused');
-        assert.notOk(dataGrid.hasEditData(), 'changes are saved');
-    });
-
-    // T553067
-    QUnit.testInActiveWindow('Enter key on editor should prevent default behaviour', function(assert) {
-        if(devices.real().deviceType !== 'desktop') {
-            assert.ok(true, 'keyboard navigation is disabled for not desktop devices');
-            return;
-        }
-
-        // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [{ name: 'name 1', value: 1 }, { name: 'name 2', value: 2 }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: [{ dataField: 'name', allowEditing: false }, { dataField: 'value', showEditorAlways: true }]
-        });
-        const navigationController = dataGrid.getController('keyboardNavigation');
-
-        this.clock.tick();
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-        $(':focus').on('focusout', function(e) {
-        // emulate browser behaviour
-            $(e.target).trigger('change');
-        });
-        $(':focus').val('test');
-
-        // act
-        const event = $.Event('keydown', { target: $(':focus').get(0) });
-        navigationController._keyDownHandler({ key: 'Enter', keyName: 'enter', originalEvent: event });
-        this.clock.tick();
-
-        // assert
-        assert.ok(event.isDefaultPrevented(), 'keydown event is prevented');
-        assert.equal(dataGrid.cellValue(0, 0), 'test', 'cell value is changed');
-    });
-
-    // T816039
-    QUnit.testInActiveWindow('Focus should be correct after change value and click to another row if showEditorAlways is true', function(assert) {
-    // arrange
-        const dataSource = [{ id: 1, name: 'name 1' }, { id: 2, name: 'name 2' }];
-        const dataGrid = createDataGrid({
-            dataSource: dataSource,
-            keyExpr: 'id',
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: [{ dataField: 'name', showEditorAlways: true }]
-        });
-
-        this.clock.tick();
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-        $(':focus').on('focusout', function(e) {
-        // emulate browser behaviour
-            $(e.target).trigger('change');
-        });
-
-        // act
-        $(':focus').val('test');
-        const $secondRowEditor = $(dataGrid.getRowElement(1)).find('.dx-texteditor-input');
-        $secondRowEditor.trigger('dxpointerdown');
-        $secondRowEditor.trigger('focus');
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataSource[0].name, 'test', 'data is changed');
-        assert.equal($(dataGrid.getRowElement(1)).find(':focus').length, 1, 'focus in second row');
-    });
-
-    QUnit.testInActiveWindow('Datebox editor\'s enter key handler should be replaced by noop (T819067)', function(assert) {
-        if(devices.real().deviceType !== 'desktop') {
-            assert.ok(true, 'keyboard navigation is disabled for not desktop devices');
-            return;
-        }
-
-        // arrange
-        const rowsViewWrapper = dataGridWrapper.rowsView;
-        const dataGrid = createDataGrid({
-            dataSource: [{ dateField: '2000/01/01 12:42' }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: [{
-                dataField: 'dateField',
-                dataType: 'date'
-            }]
-        });
-        this.clock.tick();
-
-        // act
-        $(dataGrid.getCellElement(0, 0)).trigger('dxclick');
-
-        // assert
-        const editor = rowsViewWrapper.getDataRow(0).getCell(0).getEditor();
-        const dateBox = editor.getElement().dxDateBox('instance');
-        const enterKeyHandler = dateBox._supportedKeys().enter;
-        assert.strictEqual(enterKeyHandler(), true, 'dateBox enter key handler is replaced');
-    });
-
-    QUnit.testInActiveWindow('Datebox editor\'s value should be selected from calendar by keyboard (T848039)', function(assert) {
-        if(devices.real().deviceType !== 'desktop') {
-            assert.ok(true, 'keyboard navigation is disabled for not desktop devices');
-            return;
-        }
-
-        [true, false].forEach(useMaskBehavior => {
-        // arrange
-            const rowsViewWrapper = dataGridWrapper.rowsView;
-            const dataGrid = createDataGrid({
-                dataSource: [{ dateField: '01/01/2000' }],
-                editing: {
-                    mode: 'cell',
-                    allowUpdating: true
-                },
-                columns: [{
-                    dataField: 'dateField',
-                    dataType: 'date',
-                    editorOptions: {
-                        useMaskBehavior: useMaskBehavior
-                    }
-                }]
-            });
-            this.clock.tick();
-
-            // act
-            dataGrid.editCell(0, 0);
-            this.clock.tick();
-
-            let editor = rowsViewWrapper.getDataRow(0).getCell(0).getEditor();
-            const instance = editor.getElement().dxDateBox('instance');
-            const keyboard = keyboardMock(editor.getInputElement());
-
-            instance.open();
-            keyboard
-                .keyDown('left')
-                .press('enter');
-
-            // assert
-            editor = rowsViewWrapper.getDataRow(0).getCell(0).getEditor();
-            assert.equal(editor.getInputElement().val(), '12/31/1999', `dateBox value is changed if useMaskBehavior is ${useMaskBehavior}`);
-        });
-    });
-
-    QUnit.testInActiveWindow('dataGrid resize generates exception if fixed column presents and validation applied in cell edit mode (T629168)', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ Test: 'a', c1: 'b' }, { Test: 'c', c1: 'd' }],
-            showColumnHeaders: false,
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            columns: [
-                {
-                    dataField: 'Test',
-                    fixed: true,
-                    validationRules: [{ type: 'required' }]
-                }, 'c1'
-            ]
-        });
-        const that = this;
-
-        that.clock.tick();
-
-        // act
-        dataGrid.cellValue(0, 0, '');
-
-        that.clock.tick();
-
-        $(dataGrid.getCellElement(0, 0)).trigger('dxclick');
-
-        that.clock.tick();
-
-        dataGrid.updateDimensions();
-
-        // assert
-        assert.ok(true, 'no exceptions');
     });
 
     QUnit.test('expandAll', function(assert) {
@@ -11922,87 +10025,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(dataGrid.getVisibleRows()[0].data.id, 101, 'first visible row is changed');
     });
 
-    // T551304
-    QUnit.test('row should rendered after editing if scrolling mode is virtual', function(assert) {
-    // arrange, act
-
-        const array = [];
-        for(let i = 0; i < 4; i++) {
-            array.push({ id: i, text: 'text ' + i });
-        }
-
-        const dataGrid = createDataGrid({
-            scrolling: {
-                mode: 'virtual'
-            },
-            paging: {
-                pageSize: 2
-            },
-            dataSource: array
-        });
-
-        this.clock.tick();
-
-        // act
-        dataGrid.cellValue(2, 1, 666);
-        dataGrid.saveEditData();
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataGrid.getVisibleRows().length, 4, 'visible row count');
-        assert.equal(dataGrid.cellValue(2, 1), 666, 'value is changed');
-        assert.equal(dataGrid.hasEditData(), false, 'no unsaved data');
-    });
-
-    // T837104
-    QUnit.test('Update should work after scrolling if scrolling mode is infinite and refresh mode is repaint', function(assert) {
-        const dataGrid = createDataGrid({
-            height: 100,
-            loadingTimeout: undefined,
-            remoteOperations: true,
-            dataSource: {
-                key: 'id',
-                load(options) {
-                    const items = [];
-
-                    for(let i = options.skip; i < options.skip + options.take; i++) {
-                        const id = i + 1;
-                        items.push({ id: id, name: 'test ' + id });
-                    }
-
-                    return items;
-                },
-                update() {
-                }
-            },
-            paging: {
-                pageSize: 5
-            },
-            scrolling: {
-                mode: 'infinite',
-                useNative: false
-            },
-            editing: {
-                allowUpdating: true,
-                refreshMode: 'repaint'
-            }
-        });
-
-        // act
-        dataGrid.cellValue(0, 'name', 'updated');
-        dataGrid.saveEditData();
-        dataGrid.getScrollable().scrollTo({ top: 10000 });
-        dataGrid.getScrollable().scrollTo({ top: 10000 });
-        dataGrid.cellValue(9, 'name', 'updated');
-        dataGrid.saveEditData();
-
-        // assert
-        assert.equal(dataGrid.getVisibleRows().length, 15, 'visible row count');
-        assert.deepEqual(dataGrid.getVisibleRows()[0].data, { id: 1, name: 'updated' }, 'row 1 is updated');
-        assert.deepEqual(dataGrid.getVisibleRows()[1].data, { id: 2, name: 'test 2' }, 'row 2 is not updated');
-        assert.deepEqual(dataGrid.getVisibleRows()[9].data, { id: 10, name: 'updated' }, 'row 10 is updated');
-    });
-
     QUnit.test('Duplicate rows should not be rendered if virtual scrolling enabled and column has values on second page only', function(assert) {
     // arrange, act
 
@@ -12304,40 +10326,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.ok(!errorMessage, 'There is no errors');
     });
 
-    QUnit.test('Create new row when grouping and group summary (T644293)', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columns: [
-                'field1',
-                {
-                    dataField: 'field2',
-                    groupIndex: 0
-                }
-            ],
-            dataSource: {
-                store: [{ field1: 1, field2: 2 }, { field1: 3, field2: 4 }]
-            },
-            summary: {
-                groupItems: [
-                    {
-                        column: 'field1',
-                        summaryType: 'count',
-                        showInGroupFooter: true
-                    }
-                ]
-            }
-        });
-
-        // act
-        dataGrid.addRow();
-        const $insertedRow = dataGrid.getVisibleRows()[0];
-
-        // assert
-        assert.equal($insertedRow.rowType, 'data', 'inserted row has the \'data\' type');
-        assert.equal($insertedRow.isNewRow, true, 'inserted row is presents and has 0 index');
-    });
-
     QUnit.test('Click on detail cell with cellIndex more than number of parent grid columns', function(assert) {
     // arrange
         const dataGrid = createDataGrid({
@@ -12571,41 +10559,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal($nestedRows.eq(0).height(), $nestedRows.eq(1).height(), 'nested row heights are synchronized after collapse');
     });
 
-    // T804060
-    QUnit.test('contentReady event should be fired after error during update', function(assert) {
-    // arrange act
-        let eventArray = [];
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columns: [{ dataField: 'id', fixed: true }, { dataField: 'name' }],
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            dataSource: {
-                load: function() {
-                    return [{ id: 1, name: 'test' }];
-                },
-                update: function() {
-                    return $.Deferred().reject('Update error');
-                }
-            },
-            onDataErrorOccurred: () => eventArray.push('onDataErrorOccurred'),
-            onContentReady: () => eventArray.push('onContentReady')
-        });
-
-        dataGrid.editCell(0, 1);
-        dataGrid.cellValue(0, 1, 'updated');
-
-        eventArray = [];
-
-        // act
-        dataGrid.saveEditData();
-
-        // assert
-        assert.deepEqual(eventArray, ['onDataErrorOccurred', 'onContentReady'], 'onContentReady fired after onDataErrorOccurred');
-    });
-
     // T607490
     QUnit.test('Scrollable should be updated after expand master detail row with nested DataGrid', function(assert) {
     // arrange
@@ -12756,145 +10709,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.deepEqual(visibleWidths[1], 'auto', 'column 2 has auto width');
         assert.deepEqual(visibleWidths[2], 'adaptiveHidden', 'column 3 is hidden');
         assert.deepEqual(visibleWidths[3], 'adaptiveHidden', 'column 4 is hidden');
-    });
-
-    QUnit.testInActiveWindow('Scroll positioned correct with fixed columns and editing', function(assert) {
-        if(devices.real().deviceType !== 'desktop') {
-            assert.ok(true, 'keyboard navigation is not actual for not desktop devices');
-            return;
-        }
-
-        // arrange, act
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            columnFixing: {
-                enabled: true
-            },
-            columns: [{ dataField: 'field1', width: 200 }, { dataField: 'field2', width: 200 }, { dataField: 'field3', width: 200 }, { dataField: 'fixedField', width: '200px', fixed: true, fixedPosition: 'right' }],
-            dataSource: {
-                store: [
-                    { field1: 1, field2: 2, field3: 3, fixedField: 4 },
-                    { field1: 4, field2: 5, field3: 3, fixedField: 6 }
-                ]
-            },
-            editing: {
-                allowUpdating: true,
-                mode: 'batch'
-            },
-            width: 400
-        });
-        const triggerTabPress = function($target) {
-            dataGrid.getController('keyboardNavigation')._keyDownHandler({
-                key: 'Tab',
-                keyName: 'tab',
-                originalEvent: {
-                    target: $target,
-                    preventDefault: commonUtils.noop,
-                    stopPropagation: commonUtils.noop,
-                    isDefaultPrevented: function() { return false; }
-                }
-            }, true);
-        };
-
-        // act
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-
-        triggerTabPress(dataGrid.getCellElement(0, 0));
-        this.clock.tick();
-
-        triggerTabPress(dataGrid.getCellElement(0, 1));
-        this.clock.tick();
-
-        // assert
-        assert.equal(dataGrid.getView('rowsView').getScrollable().scrollLeft(), 400, 'Correct offset');
-    });
-
-    // T532658
-    QUnit.test('Cancel editing should works correctly if editing mode is form and masterDetail row is shown', function(assert) {
-    // arrange
-        const items = [{ firstName: 'Alex', lastName: 'Black' }, { firstName: 'John', lastName: 'Dow' }];
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'form',
-                allowUpdating: true
-            },
-            dataSource: items,
-            columns: ['firstName', 'lastName']
-        });
-
-        dataGrid.expandRow(items[0]);
-        dataGrid.editRow(0);
-
-        assert.ok($(dataGrid.getRowElement(0)).hasClass('dx-datagrid-edit-form'), 'row 0 is edit form row');
-        assert.ok(dataGrid.getVisibleRows()[0].isEditing, 'row 0 isEditing');
-
-        // act
-        dataGrid.cancelEditData();
-
-        // assert
-        assert.ok($(dataGrid.getRowElement(0)).hasClass('dx-data-row'), 'row 0 is data row');
-        assert.notOk(dataGrid.getVisibleRows()[0].isEditing, 'row 0 isEditing');
-
-        assert.ok($(dataGrid.getRowElement(1)).hasClass('dx-master-detail-row'), 'row 1 is master detail row');
-        assert.notOk($(dataGrid.getRowElement(1)).hasClass('dx-datagrid-edit-form'), 'row 1 is not edit form row');
-        assert.notOk(dataGrid.getVisibleRows()[1].isEditing, 'row 1 isEditing');
-
-        assert.ok($(dataGrid.getRowElement(2)).hasClass('dx-data-row'), 'row 2 is data row');
-    });
-
-    // T736360
-    QUnit.test('Editing should be started without errors if update form items in contentReady', function(assert) {
-    // arrange
-        const items = [{ firstName: 'Alex', lastName: 'Black' }, { firstName: 'John', lastName: 'Dow' }];
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            editing: {
-                mode: 'form',
-                allowUpdating: true,
-                form: {
-                    items: [{
-                        itemType: 'tabbed',
-                        tabs: [{
-                            title: 'First Name',
-                            items: [{
-                                dataField: 'firstName'
-                            }]
-                        }, {
-                            title: 'Last Name',
-                            items: [{
-                                dataField: 'lastName'
-                            }]
-                        }]
-                    }]
-                }
-            },
-            dataSource: items,
-            columns: [{
-                dataField: 'firstName',
-                validationRules: [{ type: 'required' }]
-            }, {
-                dataField: 'lastName',
-                validationRules: [{ type: 'required' }]
-            }],
-            onContentReady: function(e) {
-                const $tabPanel = $(e.element).find('.dx-tabpanel');
-                if($tabPanel.length) {
-                    const tabPanel = $tabPanel.dxTabPanel('instance');
-                    tabPanel.option('items', tabPanel.option('items'));
-                    tabPanel.option('selectedIndex', 1);
-                }
-            }
-        });
-
-        // act
-        dataGrid.editRow(0);
-
-        assert.ok($(dataGrid.getRowElement(0)).hasClass('dx-datagrid-edit-form'), 'row 0 is edit form row');
-        assert.ok(dataGrid.getVisibleRows()[0].isEditing, 'row 0 isEditing');
     });
 
     [true, false].forEach(useLegacyKeyboardNavigation => {
@@ -13696,32 +11510,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(dataGrid.cellValue(1, 0), 1, 'second row value');
     });
 
-    // T394873
-    QUnit.test('Column widths must be kept after cell edit', function(assert) {
-    // arrange
-        const $grid = $('#dataGrid').dxDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ name: 'James Bond', code: '007' }],
-            columnAutoWidth: true,
-            editing: {
-                allowUpdating: true,
-                mode: 'batch'
-            }
-        });
-        const gridInstance = $grid.dxDataGrid('instance');
-
-        const visibleWidths = [gridInstance.columnOption(0, 'visibleWidth'), gridInstance.columnOption(1, 'visibleWidth')];
-
-        // act
-        gridInstance.editCell(0, 0);
-
-        // assert
-        const newVisibleWidths = [gridInstance.columnOption(0, 'visibleWidth'), gridInstance.columnOption(1, 'visibleWidth')];
-        assert.equal($grid.find('input').length, 1, 'one editor is rendered');
-
-        assert.deepEqual(newVisibleWidths, visibleWidths, 'visibleWidths are not changed');
-    });
-
     QUnit.test('Repaint row', function(assert) {
     // arrange
         const dataSource = new DataSource({
@@ -14322,163 +12110,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.strictEqual($(dataGrid.getRowElement(1)).children().length, 4, 'data cell count for second data row is correct');
     });
 
-    // T757163
-    QUnit.test('cancelEditData in onRowUpdating event for boolean column if repaintChangesOnly is true', function(assert) {
-    // arrange
-        let rowUpdatingCallCount = 0;
-        const dataGrid = createDataGrid({
-            dataSource: [
-                { id: 1, value: true },
-                { id: 2, value: true }
-            ],
-            keyExpr: 'id',
-            loadingTimeout: undefined,
-            repaintChangesOnly: true,
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            onRowUpdating: function(e) {
-                rowUpdatingCallCount++;
-                if(e.key === 1) {
-                    e.cancel = true;
-                    e.component.cancelEditData();
-                }
-            }
-        });
-
-        const $firstCheckBoxCell = $(dataGrid.getCellElement(0, 1));
-        const $secondCheckBoxCell = $(dataGrid.getCellElement(1, 1));
-
-        // act
-        $firstCheckBoxCell.find('.dx-checkbox').dxCheckBox('instance').option('value', false);
-
-        // assert
-        assert.strictEqual(rowUpdatingCallCount, 1, 'onRowUpdating is called');
-        assert.strictEqual($(dataGrid.getCellElement(0, 1)).find('.dx-checkbox').dxCheckBox('instance').option('value'), true, 'first checkbox value is canceled');
-        assert.notStrictEqual($(dataGrid.getCellElement(0, 1)).get(0), $firstCheckBoxCell.get(0), 'first checkbox cell is changed');
-        assert.strictEqual($(dataGrid.getCellElement(1, 1)).get(0), $secondCheckBoxCell.get(0), 'second checkbox cell is not changed');
-    });
-
-    QUnit.test('DataGrid should repaint editors on cancelEditData method if repaintChangesOnly is true (T820847)', function(assert) {
-    // arrange
-        const rowsViewWrapper = dataGridWrapper.rowsView;
-        const dataGrid = createDataGrid({
-            dataSource: [{ id: 1 }, { id: 2 }],
-            repaintChangesOnly: true,
-            editing: {
-                mode: 'cell',
-                allowUpdating: true
-            },
-            loadingTimeout: undefined
-        });
-
-        // act
-        dataGrid.editCell(0, 0);
-        // assert
-        let editor = rowsViewWrapper.getDataRow(0).getCell(0).getEditor();
-        assert.ok(editor.getInputElement().length > 0, 'cell has editor');
-
-        // act
-        dataGrid.cancelEditData();
-        // assert
-        editor = rowsViewWrapper.getDataRow(0).getCell(0).getEditor();
-        assert.equal(editor.getInputElement().length, 0, 'cell has no editor');
-    });
-
-    QUnit.test('Using watch in cellPrepared event for editor if repaintChangesOnly', function(assert) {
-    // arrange
-        const dataSource = new DataSource({
-            store: {
-                type: 'array',
-                key: 'id',
-                data: [
-                    { id: 1, field1: 'test1' },
-                    { id: 2, field1: 'test2' },
-                    { id: 3, field1: 'test3' },
-                    { id: 4, field1: 'test4' }
-                ]
-            }
-        });
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: dataSource,
-            columns: ['id', 'field1'],
-            editing: {
-                mode: 'cell'
-            },
-            repaintChangesOnly: true,
-            onCellPrepared: function(e) {
-                if(e.isEditing) {
-                    e.watch(function() {
-                        return e.column.calculateCellValue(e.data);
-                    }, function() {
-                        $(e.cellElement).addClass('changed');
-                    });
-                }
-            }
-        });
-
-        this.clock.tick();
-        dataGrid.editCell(0, 1);
-
-        dataSource.store().update(1, { field1: 'test5' });
-
-        // assert
-        const $cell = $(dataGrid.getCellElement(0, 1));
-
-        // act
-        dataGrid.refresh(true);
-        this.clock.tick();
-
-        // assert
-        assert.ok($(dataGrid.getCellElement(0, 1)).is($cell), 'first cell isn\'t updated');
-        assert.ok($cell.hasClass('changed'), 'class changed is added');
-        assert.equal($(dataGrid.element()).find('.changed').length, 1, 'class changed is added to one cell only');
-    });
-
-    QUnit.test('watch in cellPrepared should works after cell editing', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [
-                { id: 1, field1: 'test1' },
-                { id: 2, field1: 'test2' }
-            ],
-            keyExpr: 'id',
-            loadingTimeout: undefined,
-            repaintChangesOnly: true,
-            editing: {
-                mode: 'cell'
-            },
-            onCellPrepared: function(e) {
-                if(e.rowType === 'data') {
-                    e.watch(function() {
-                        return e.key === activeRowKey;
-                    }, function(isActive) {
-                        $(e.cellElement).toggleClass('active', isActive);
-                    });
-                }
-            },
-            columns: ['id', 'field1']
-        });
-
-        this.clock.tick();
-
-        dataGrid.editCell(0, 1);
-        dataGrid.closeEditCell();
-
-        this.clock.tick();
-
-        // act
-        const activeRowKey = 1;
-        dataGrid.refresh(true);
-
-        // assert
-        assert.ok($(dataGrid.getCellElement(0, 0)).hasClass('active'), 'active class is added to first cell');
-        assert.ok($(dataGrid.getCellElement(0, 1)).hasClass('active'), 'active class is added to second cell');
-        assert.notOk($(dataGrid.getCellElement(1, 0)).hasClass('active'), 'active class is not added to second row');
-    });
-
     QUnit.test('watch in cellPrepared should works after push', function(assert) {
     // arrange
         const dataGrid = createDataGrid({
@@ -14591,141 +12222,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.equal(dataGrid.getVisibleColumns()[0].visibleWidth, 30, 'visibleWidth for first groupExpand column');
     });
 
-    QUnit.test('Stop watch in cellPrepared event for editor if repaintChangesOnly', function(assert) {
-    // arrange
-        const dataSource = new DataSource({
-            store: {
-                type: 'array',
-                key: 'id',
-                data: [
-                    { id: 1, field1: 'test1' },
-                    { id: 2, field1: 'test2' },
-                    { id: 3, field1: 'test3' },
-                    { id: 4, field1: 'test4' }
-                ]
-            }
-        });
-        let watchUpdateCount = 0;
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: dataSource,
-            columns: ['id', 'field1'],
-            editing: {
-                mode: 'cell'
-            },
-            repaintChangesOnly: true,
-            onCellPrepared: function(e) {
-                if(e.isEditing) {
-                    const stopWatch = e.watch(function() {
-                        return e.column.calculateCellValue(e.data);
-                    }, function() {
-                        watchUpdateCount++;
-                        if(watchUpdateCount === 2) {
-                            stopWatch();
-                        }
-                    });
-                }
-            }
-        });
-
-        this.clock.tick();
-        dataGrid.editCell(0, 1);
-
-        for(let i = 0; i < 5; i++) {
-            dataSource.store().update(1, { field1: 'changed' + i });
-            dataGrid.refresh(true);
-            this.clock.tick();
-        }
-
-        // assert
-        assert.equal(watchUpdateCount, 2, 'watch is stopped on second update');
-    });
-
-    QUnit.test('Using watch in masterDetail template if repaintChangesOnly', function(assert) {
-    // arrange
-        const dataSource = new DataSource({
-            store: {
-                type: 'array',
-                key: 'id',
-                data: [
-                    { id: 1, field1: 'test1' },
-                    { id: 2, field1: 'test2' },
-                    { id: 3, field1: 'test3' },
-                    { id: 4, field1: 'test4' }
-                ]
-            }
-        });
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: dataSource,
-            columns: ['id', 'field1'],
-            editing: {
-                mode: 'cell'
-            },
-            repaintChangesOnly: true,
-            masterDetail: {
-                template: function(container, options) {
-                    const $detail = $('<div>').addClass('detail').appendTo(container);
-
-                    $detail.text(options.data.field1);
-
-                    options.watch(function(data) {
-                        return data.field1;
-                    }, function(newValue) {
-                        $detail.text(newValue);
-                    });
-                }
-            }
-        });
-
-        this.clock.tick();
-        dataGrid.expandRow(1);
-
-        dataSource.store().update(1, { field1: 'changed' });
-
-        // assert
-        const $detail = $(dataGrid.element()).find('.detail');
-
-        // act
-        dataGrid.refresh(true);
-        this.clock.tick();
-
-        // assert
-        assert.ok($(dataGrid.element()).find('.detail').is($detail), 'detail element isn\'t updated');
-        assert.strictEqual($detail.text(), 'changed', 'detail text is changed');
-    });
-
-    // T800483
-    QUnit.test('No error after detail collapse and popup editing form closing if repaintChangesOnly is true', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            repaintChangesOnly: true,
-            dataSource: [{
-                'Id': 1,
-                'CompanyName': 'Super Mart of the West'
-            }],
-            keyExpr: 'Id',
-            columns: ['CompanyName'],
-            masterDetail: {
-                enabled: true,
-            },
-            editing: {
-                mode: 'popup',
-                allowUpdating: true,
-            }
-        });
-
-        // act
-        dataGrid.expandRow(1);
-        dataGrid.collapseRow(1);
-        dataGrid.editRow(0);
-        dataGrid.cancelEditData();
-
-        // assert
-        assert.notOk($('.dx-datagrid-edit-popup').is(':visible'), 'editor popup is hidden');
-    });
-
     QUnit.test('push changes for adaptive row', function(assert) {
     // arrange
         const dataSource = new DataSource({
@@ -14803,39 +12299,6 @@ QUnit.module('API methods', baseModuleConfig, () => {
         assert.ok($updatedCellElements.eq(0).is($cellElements.eq(0)), 'first cell isn\'t changed');
         assert.notOk($updatedCellElements.eq(1).is($cellElements.eq(1)), 'second cell is changed');
         assert.strictEqual($updatedCellElements.eq(1).text(), 'Sum: 500', 'cell value is updated');
-    });
-
-    // T851082
-    QUnit.test('Row deleting should works if recalculateWhileEditing is enabled and refreshMode is repaint', function(assert) {
-    // arrange
-        const dataGrid = createDataGrid({
-            dataSource: [{ id: 1 }],
-            keyExpr: 'id',
-            editing: {
-                refreshMode: 'repaint',
-                mode: 'batch',
-                allowDeleting: true
-            },
-            summary: {
-                recalculateWhileEditing: true,
-                totalItems: [{
-                    column: 'id',
-                    summaryType: 'count'
-                }]
-            }
-        });
-        this.clock.tick();
-
-        // act
-        dataGrid.deleteRow(0);
-        this.clock.tick();
-
-        dataGrid.saveEditData();
-        this.clock.tick();
-
-        // assert
-        assert.strictEqual(dataGrid.getVisibleRows().length, 0, 'row is removed');
-        assert.strictEqual(dataGrid.getTotalSummaryValue('id'), 0, 'summary is updated');
     });
 
     QUnit.test('Refresh with changesOnly for fixed columns', function(assert) {
@@ -16468,48 +13931,6 @@ QUnit.module('Formatting', baseModuleConfig, () => {
             }
         }), '216 rub');
     });
-
-    QUnit.testInActiveWindow('Validation message should be positioned relative cell in material theme', function(assert) {
-    // arrange
-        let overlayTarget;
-        const origIsMaterial = themes.isMaterial;
-
-        const dataGrid = createDataGrid({
-            loadingTimeout: undefined,
-            dataSource: [{ Test: '' }],
-            editing: {
-                mode: 'batch',
-                allowUpdating: true
-            },
-            columns: [{
-                dataField: 'Test',
-                validationRules: [{ type: 'required' }]
-            }]
-        });
-
-        // act
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-
-        // assert
-        overlayTarget = dataGrid.$element().find('.dx-invalid-message').data('dxOverlay').option('target');
-        assert.ok(overlayTarget.hasClass('dx-highlight-outline'), 'target in generic theme');
-
-        // act
-        dataGrid.closeEditCell();
-        this.clock.tick();
-
-        themes.isMaterial = function() { return true; };
-
-        dataGrid.editCell(0, 0);
-        this.clock.tick();
-
-        // assert
-        overlayTarget = dataGrid.$element().find('.dx-invalid-message').data('dxOverlay').option('target');
-        assert.ok(overlayTarget.hasClass('dx-editor-cell'), 'target in material theme');
-
-        themes.isMaterial = origIsMaterial;
-    });
 });
 
 QUnit.module('Row dragging', baseModuleConfig, () => {
@@ -16984,39 +14405,6 @@ QUnit.module('Column Resizing', baseModuleConfig, () => {
         assert.equal($dataGrid.children().height(), 300, 'widget\'s height is not changed');
         assert.equal(instance.columnOption(0, 'width'), 30, 'column 0 width');
         assert.equal(instance.columnOption(1, 'width'), 170, 'column 1 width');
-    });
-
-    QUnit.test('Resize is not called after editCell', function(assert) {
-        // arrange
-        const dataGrid = $('#dataGrid').dxDataGrid({
-            dataSource: {
-                store: [
-                    { firstName: 1, lastName: 2, room: 3, birthDay: 4 },
-                    { firstName: 4, lastName: 5, room: 3, birthDay: 6 }
-                ]
-            },
-            editing: {
-                allowUpdating: true,
-                mode: 'batch'
-            }
-        }).dxDataGrid('instance');
-
-        const resizingController = dataGrid.getController('resizing');
-        const rowsView = dataGrid.getView('rowsView');
-
-        sinon.spy(resizingController, 'resize');
-        sinon.spy(rowsView, 'synchronizeRows');
-        this.clock.tick();
-        assert.equal(resizingController.resize.callCount, 1, 'resize call count before editCell');
-        assert.equal(rowsView.synchronizeRows.callCount, 1, 'synchronizeRows call count before editCell');
-
-        // act
-        dataGrid.editCell(0, 0);
-
-        // assert
-        assert.ok(dataGrid.getController('editing').isEditing());
-        assert.equal(resizingController.resize.callCount, 1, 'resize call count after editCell');
-        assert.equal(rowsView.synchronizeRows.callCount, 2, 'synchronizeRows call count after editCell');
     });
 
     QUnit.test('Resize columns and move column to another position in virtual scrolling mode (T222418)', function(assert) {
